remove duplication - create queue
